requirements.txt
# Core ML/RL
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Network Analysis
networkx>=3.0

# API
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0

# Visualization (for analytics, optional)
matplotlib>=3.7.0

# Gen AI & LLM
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
typing-extensions>=4.5.0

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
